Rewrite the cache functionality
The previous method was working most of the times, but not always.
Especially during the tests with blank browser profiles, it happened
multiple times to me that the could not been read, because it didn't
exist yet.

I've rewritten the cache-based functionality to always have an empty
fallback. Furthermore, the .then() chains were rewritten to use the
try/catch pattern.